Fix a typo in the help text for --output-compression, don't allow unknown compression method arguments for that options (#105)
This PR fixes a typo in the --help text for --output-compression (which has listed gz2 as a valid argument, but the supported argument is actually gz). It also changes the behavior when an unknown argument is supplied: previously, it silently fell back to bz2, with this PR, an error is thrown.