Don't consider untagged nodes for geometric relations by default (#107)
Since #101, the geometries of untagged nodes are written by default. This adds around 10 billion triples for OSM Planet, which is fine. However, considering these nodes for the geometric relations (`contains`, `intersects`, ...) adds an enormous number of triples. It is also not necessary since the information which node is a member of which way can be obtained via #111. 

Therefore, the default is now to *not* consider untagged nodes for the geometric relations, but to output the triples that express which node is a member of which way. The triples that express which way is a member of which relation were already output by default before.

To change the new default behavior, there are now the options `--add-untagged-nodes-geometric-relations` (to consider untagged nodes for the geometric relations, despite the enormous number of triples this incurs) and `--no-member-triples` (to *not* output the triples that express which node is a member of which way and which way is a member of which relation).