Separate prefix for untagged nodes with `--iri-prefix-for-untagged-nodes` (#120)
As of this writing, the complete OSM data has 11 B objects, around 90% of which are untagged nodes, that is, nodes with zero tags (key-value pairs). Most queries involve a join of a set of tagged objects with their geometry, a minimal example being `?object osmkey:amenity "post_box" ; geo:hasGeometry ?geometry .` This is a join of a relatively small table (400 k rows in this case) with a very large table (11 B rows). In most queries, this will be a merge join, which is efficient in principle. However, the matching rows are typically "dense" in the large table because IRIs for tagged and untagged nodes both start with `https://www.openstreetmap.org/node/` followed by a number that has nothing to do with whether the node is tagged or untagged. Then the merge join has to scan through most of the very large table, which is expensive due to its sheer size.

With `--iri-prefix-for-untagged-nodes`, a distinct prefix can be given to untagged nodes, for example `http://www.openstreetmap.org/node/` (which has the nice side property of still working in a browser). The matching rows in the merge join from the example above are then all contained in a range of only around 1 B rows, which for an engine such as QLever (which is smart enough to figure out that it can skip the remaining 90% of the rows) speeds up the merge join by a factor of around 10.

When `--iri-prefix-for-untagged-nodes` is used with a prefix other than `https://www.openstreetmap.org/node/`, then also the objects of `geo:hasGeometry` and the subjects of `geo:asWKT` are given names such that the prefix for tagged and untagged nodes is different, for example, `osm2rdfgeom:osmnode_tagged_1` and `osm2rdfgeom:osmnode_untagged_137`. That way, the same speedups can be achieved for joins with the `geo:asWKT` predicate. This is important because most OSM queries involve a join with `geo:hasGeometry/geo:asWKT`, so both of the mentioned predicates.

The modified node IRIs appear in three kinds of triples in the output of `osm2rdf`. First, as subject and object of `geo:hasGeometry` . Second, as subject of `geo:asWKT`. Third, as object of `osmway:member_id`. Using the modified IRI in the first two places is trivial because when we write those triples, we have seen the node for the first time and we know whether its tagged or untagged. For the third place, it's more tricky because this happens later in the output and we need to remember whether a node is tagged or untagged. So far, only the coordinates of each node is remembered. We remember the additional bit by using the most significant bit of the `y` coordinate. That way, no more space than is used than so far, which is important because of the sheer number of nodes.

NOTE 1: To implement this, we need a modified version of the headers `Location.h` and `NodeLocationsForWays.h` from `libosmium`. According to https://github.com/osmcode/libosmium/issues/395, this change cannot be made in `libosmium` itself, at least for now. We therefore use modified copies. These potentially have to be updated when pulling newer version of `libosmium`.

NOTE 2: There is still a bug in that the precomputed geometric relations (`ogc:sfContains`, etc.) don't use the modified IRIs when tagged and untagged nodes have different prefixes. This is easy to fix and will be done in a subsequent PR.

Co-authored-by: Patrick Brosi <brosip@informatik.uni-freiburg.de>