Separate prefix for untagged nodes (#120)
Add the option to use different IRI prefixes for untagged and tagged nodes. Most queries use `geo:hasGeometry/geo:asWKT` in conjunction with other constraints on the subject (for example, all highways). These constraint usually restrict to tagged nodes, which are much fewer than untagged
nodes. By giving the untagged nodes a separate IRI prefix, the respective join in QLever is much more efficient because the block of ??? pertaining to untagged nodes can be skipped. For the way members, which do not access to the node tags, this is implemented by storing information on whether a node location was from a tagged node in the MSB of the locations `y`-coordinate.

---------

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>
Co-authored-by: Patrick Brosi <brosip@informatik.uni-freiburg.de>