Fix remaining issues regarding `--iri-prefix-for-untagged-nodes` (#124)
Fixes two issues with #120 (which introduced the option to have separate prefixes for tagged and untagged nodes), which went unnoticed during the review phase:

1. The computation phase of the geometric relations ignored the separate prefixes and always used the standard prefix
2. The separate prefixes for tagged and untagged nodes always used `osm`, even when `--source-dataset` was set to `OHM`

The fix for 1 is trivial. The fix for 2 is as follows: If `--source-dataset` is `OSM`, use `https://www.openstreetmap.org/node` as the default prefix for nodes, if it is `OHM`, use `https://www.openhistoricalmap/node`. If `--iri-prefix-for-untagged-nodes` is not set or empty, use that default prefix for both tagged and untagged nodes, otherwise use it for the tagged nodes.