Dump `osm2rdf:length` and `osm2rdf:area` as meter and m² (#127)
Changes the units of osm2rdf:length to meters, and of osm2rdf:area to square meters.

Meter lengths are computed by summing up the great-circle distances between line anchor points using the haversine formular.

Square meter areas are computed by projecting polygon coordinates using an area-preserving projection (the Lambert cylindrical equal-area projection) and computing the area using Gauss's area formula. This simple projection assumes that the earth is a perfect sphere, but the approximation error should be negligible for typical polygons in OSM. For example, the approximation error of the area of Switzerland is below 0.3%.