Separate reader for node/way/area and relation to fix missing relation geometries (#129)
Adds separate readers for nwa (node/way/area) and relation to fix an issue where way member geometries may not have been seen at the time a relation arrives at the RelationHandler when using multiple threads.

Fixes #128.

On the side, add hasCompleteGeometry true triples to relation areas for consistency (because the same triple is present for non-area relations) (fixes #130), and fix a subtle issue with the area classification for ways. Because of this issue, untagged closed ways (of which there are few, if any, in OSM) had no geometry object. Fixes #131, see there for details.